fs: Enforce POSIX directory semantics in VFS and fix write_fs return type
- Modified `read_fs` and `write_fs` in `kernel/fs/vfs.c` to explicitly
  return `-EISDIR` when attempting to read from or write to a directory node.
- Updated `write_fs` and `write_type_t` return types from `uint32_t` to
  `ssize_t` to correctly propagate negative error codes.
- Updated all filesystem driver `write` implementations (`fat32`, `jfs`,
  `devfs`, `shmfs`, `socket`, `partition`, `tty`) to match the new signature.

+++ include/fs/vfs.h
@@ -39,7 +39,7 @@ struct linux_dirent64 {
 #define DT_WHT     14
 
 typedef ssize_t (*read_type_t)(struct fs_node*, uint32_t, uint32_t, uint8_t*);
-typedef uint32_t (*write_type_t)(struct fs_node*, uint32_t, uint32_t, uint8_t*);
+typedef ssize_t (*write_type_t)(struct fs_node*, uint32_t, uint32_t, uint8_t*);
 typedef void (*open_type_t)(struct fs_node*);
 typedef void (*close_type_t)(struct fs_node*);
 typedef int (*readdir_type_t)(struct fs_node*, uint32_t, struct dirent*);
@@ -87,7 +87,7 @@ typedef struct fs_node {
  * , not file nodes.
  */
 ssize_t read_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
-uint32_t write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
+ssize_t write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
 void open_fs(fs_node_t *node, uint8_t read, uint8_t write);
 void close_fs(fs_node_t *node);
 int readdir_fs(fs_node_t *node, uint32_t index, struct dirent *entry);
