+++ kernel/fs/fat32.c
@@ -399,7 +399,7 @@ static int fat32_find_dirent_in_chain(fat32_volume_t* vol, uint32_t start_cluste
 
 // Forward declarations
 ssize_t fat32_read_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
-uint32_t fat32_write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
+ssize_t fat32_write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
 void fat32_open_fs(fs_node_t *node);
 void fat32_close_fs(fs_node_t *node);
 int fat32_readdir_fs(fs_node_t *node, uint32_t index, struct dirent *entry);
@@ -492,7 +492,7 @@ ssize_t fat32_read_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *
     return res;
 }
 
-static uint32_t fat32_write_fs_impl(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
+static ssize_t fat32_write_fs_impl(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
     fat32_volume_t* vol = (fat32_volume_t*)node->ptr;
 
     if (node->inode == 0) {
@@ -603,10 +603,10 @@ static uint32_t fat32_write_fs_impl(fs_node_t *node, uint32_t offset, uint32_t s
     return bytes_written;
 }
 
-uint32_t fat32_write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
+ssize_t fat32_write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
     fat32_volume_t* vol = (fat32_volume_t*)node->ptr;
     spin_lock(&vol->lock);
-    uint32_t res = fat32_write_fs_impl(node, offset, size, buffer);
+    ssize_t res = fat32_write_fs_impl(node, offset, size, buffer);
     spin_unlock(&vol->lock);
     return res;
 }
